+++ lib/src/components/snackbar/snackbar.dart
@@ -9,6 +9,7 @@ import 'package:angular_components/src/laminate/popup/src/popup_size_provider.da
 import 'package:angular_components/src/laminate/popup/src/popup_source.dart';
 import 'package:angular_components/src/utils/angular/properties/properties.dart';
 import 'package:angular_components/src/utils/disposer/disposer.dart';
+import 'package:angular_components/src/components/material_popup/material_popup.dart';
 
 @Injectable()
 class DocumentPopupSource extends ElementPopupSource {
@@ -53,7 +54,7 @@ class DocumentPopupSource extends ElementPopupSource {
     selector: 'skawa-snackbar',
     templateUrl: 'snackbar.html',
     styleUrls: const ['snackbar.css'],
-    directives: const [PopupComponent, MaterialButtonComponent, NgIf, NgClass],
+    directives: const [MaterialPopupComponent, MaterialButtonComponent, NgIf, NgClass],
     providers: const [
       const Provider(PopupSource, useClass: DocumentPopupSource),
       const Provider(PopupSizeProvider, useFactory: sizeProviderFactory)
@@ -71,6 +72,9 @@ class SkawaSnackbarComponent implements OnDestroy {
     _disposer..addEventSink(_toggleController)..addEventSink(_triggerController);
   }
 
+  @ViewChild(MaterialPopupComponent)
+  MaterialPopupComponent popupComponent;
+
   @Input()
   String actionLabel;
 
@@ -86,21 +90,40 @@ class SkawaSnackbarComponent implements OnDestroy {
   @Input()
   set visible(val) {
     displayPopup = getBool(val);
+    if (displayPopup) {
+      popupComponent.open();
+    } else {
+      popupComponent.close();
+    }
     _cd.markForCheck();
   }
 
   get visible => displayPopup;
 
   bool displayPopup = false;
+  bool isFixed = false;
 
   bool get useAction => actionLabel != null;
   bool showAnimation = false;
 
+  DivElement parentDiv;
+  bool first = true;
+
   void visibleChanged(bool visible) {
+    if (first) {
+      parentDiv = document.body.querySelector('div[pane-id="${popupComponent.resolvedPopupRef.uniqueId}"]');
+      _disposer.addStreamSubscription(popupComponent.onOpened.listen((_) {
+        parentDiv.style.setProperty("transform", "translateX(30px) translateY(${window.screen.height - 130}px)");
+      }));
+
+      first = false;
+    }
+    if (!isFixed) {
+      parentDiv.style.position = 'fixed';
+      isFixed = true;
+    }
     showAnimation = displayPopup;
-    displayPopup = visible;
     _cd.markForCheck();
-    _toggleController.add(visible);
   }
 
   void trigger(event) {